Extract router and useRecordApi into separate modules.
