+++ lib/js/src/ReasonReact.js
@@ -2,7 +2,6 @@
 
 var Curry = require("bs-platform/lib/js/curry.js");
 var React = require("react");
-var Caml_obj = require("bs-platform/lib/js/caml_obj.js");
 var Js_primitive = require("bs-platform/lib/js/js_primitive.js");
 var Caml_builtin_exceptions = require("bs-platform/lib/js/caml_builtin_exceptions.js");
 var ReasonReactOptimizedCreateClass = require("./ReasonReactOptimizedCreateClass.js");
@@ -404,268 +403,7 @@ function wrapJsForReason(reactClass, props, children) {
         ];
 }
 
-function safeMakeEvent(eventName) {
-  if (typeof Event === "function") {
-    return new Event(eventName);
-  } else {
-    var $$event = document.createEvent("Event");
-    $$event.initEvent(eventName, true, true);
-    return $$event;
-  }
-}
-
-function path() {
-  var match = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined) {
-    var raw = match.location.pathname;
-    switch (raw) {
-      case "" : 
-      case "/" : 
-          return /* [] */0;
-      default:
-        var raw$1 = raw.slice(1);
-        var match$1 = raw$1[raw$1.length - 1 | 0];
-        var raw$2 = match$1 === "/" ? raw$1.slice(0, -1) : raw$1;
-        var a = raw$2.split("/");
-        var _i = a.length - 1 | 0;
-        var _res = /* [] */0;
-        while(true) {
-          var res = _res;
-          var i = _i;
-          if (i < 0) {
-            return res;
-          } else {
-            _res = /* :: */[
-              a[i],
-              res
-            ];
-            _i = i - 1 | 0;
-            continue ;
-          }
-        };
-    }
-  } else {
-    return /* [] */0;
-  }
-}
-
-function hash() {
-  var match = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined) {
-    var raw = match.location.hash;
-    switch (raw) {
-      case "" : 
-      case "#" : 
-          return "";
-      default:
-        return raw.slice(1);
-    }
-  } else {
-    return "";
-  }
-}
-
-function search() {
-  var match = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined) {
-    var raw = match.location.search;
-    switch (raw) {
-      case "" : 
-      case "?" : 
-          return "";
-      default:
-        return raw.slice(1);
-    }
-  } else {
-    return "";
-  }
-}
-
-function push(path) {
-  var match = typeof (history) === "undefined" ? undefined : (history);
-  var match$1 = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined && match$1 !== undefined) {
-    match.pushState(null, "", path);
-    match$1.dispatchEvent(safeMakeEvent("popstate"));
-    return /* () */0;
-  } else {
-    return /* () */0;
-  }
-}
-
-function replace(path) {
-  var match = typeof (history) === "undefined" ? undefined : (history);
-  var match$1 = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined && match$1 !== undefined) {
-    match.replaceState(null, "", path);
-    match$1.dispatchEvent(safeMakeEvent("popstate"));
-    return /* () */0;
-  } else {
-    return /* () */0;
-  }
-}
-
-function url() {
-  return /* record */[
-          /* path */path(/* () */0),
-          /* hash */hash(/* () */0),
-          /* search */search(/* () */0)
-        ];
-}
-
-function watchUrl(callback) {
-  var match = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined) {
-    var watcherID = function () {
-      return Curry._1(callback, url(/* () */0));
-    };
-    match.addEventListener("popstate", watcherID);
-    return watcherID;
-  } else {
-    return (function () {
-        return /* () */0;
-      });
-  }
-}
-
-function unwatchUrl(watcherID) {
-  var match = typeof (window) === "undefined" ? undefined : (window);
-  if (match !== undefined) {
-    match.removeEventListener("popstate", watcherID);
-    return /* () */0;
-  } else {
-    return /* () */0;
-  }
-}
-
-function useUrl() {
-  var match = React.useState(url);
-  var setUrl = match[1];
-  React.useLayoutEffect((function () {
-          var watcherId = watchUrl((function (url) {
-                  return Curry._1(setUrl, (function () {
-                                return url;
-                              }));
-                }));
-          return (function () {
-                    return unwatchUrl(watcherId);
-                  });
-        }), ([]));
-  return match[0];
-}
-
-function useRecordApi(componentSpec) {
-  var initialState = React.useMemo(componentSpec[/* initialState */10], ([]));
-  var unmountSideEffects = React.useRef(/* array */[]);
-  var match = React.useReducer((function (fullState, action) {
-          var recordApiStateReturn = Curry._2(componentSpec[/* reducer */12], action, fullState[/* state */1]);
-          if (typeof recordApiStateReturn === "number") {
-            return fullState;
-          } else {
-            switch (recordApiStateReturn.tag | 0) {
-              case 0 : 
-                  return /* record */[
-                          /* sideEffects : array */[],
-                          /* state */recordApiStateReturn[0]
-                        ];
-              case 1 : 
-                  return /* record */[
-                          /* sideEffects *//* array */[recordApiStateReturn[0]].concat(fullState[/* sideEffects */0]),
-                          /* state */fullState[/* state */1]
-                        ];
-              case 2 : 
-                  return /* record */[
-                          /* sideEffects *//* array */[recordApiStateReturn[1]].concat(fullState[/* sideEffects */0]),
-                          /* state */recordApiStateReturn[0]
-                        ];
-              
-            }
-          }
-        }), /* record */[
-        /* sideEffects : array */[],
-        /* state */initialState
-      ]);
-  var send = match[1];
-  var match$1 = match[0];
-  var sideEffects = match$1[/* sideEffects */0];
-  var self = [];
-  Caml_obj.caml_update_dummy(self, /* record */[
-        /* handle */(function (fn, payload) {
-            return Curry._2(fn, payload, self);
-          }),
-        /* state */match$1[/* state */1],
-        /* retainedProps */componentSpec[/* retainedProps */11],
-        /* send */send,
-        /* onUnmount */(function (sideEffect) {
-            unmountSideEffects.current.push(sideEffect);
-            return /* () */0;
-          })
-      ]);
-  var state = Curry._1(componentSpec[/* willReceiveProps */3], self);
-  var self$1 = [];
-  Caml_obj.caml_update_dummy(self$1, /* record */[
-        /* handle */(function (fn, payload) {
-            return Curry._2(fn, payload, self$1);
-          }),
-        /* state */state,
-        /* retainedProps */componentSpec[/* retainedProps */11],
-        /* send */send,
-        /* onUnmount */(function (sideEffect) {
-            unmountSideEffects.current.push(sideEffect);
-            return /* () */0;
-          })
-      ]);
-  var oldSelf = React.useRef(self$1);
-  var hasBeenCalled = React.useRef(false);
-  React.useLayoutEffect((function () {
-          Curry._1(componentSpec[/* didMount */4], self$1);
-          return (function () {
-                    unmountSideEffects.current.forEach((function (fn) {
-                            return Curry._1(fn, /* () */0);
-                          }));
-                    unmountSideEffects.current = /* array */[];
-                    return Curry._1(componentSpec[/* willUnmount */6], self$1);
-                  });
-        }), ([]));
-  React.useLayoutEffect((function () {
-          if (hasBeenCalled.current) {
-            hasBeenCalled.current = true;
-          } else {
-            Curry._1(componentSpec[/* didUpdate */5], /* record */[
-                  /* oldSelf */oldSelf.current,
-                  /* newSelf */self$1
-                ]);
-          }
-          sideEffects.forEach((function (fn) {
-                  return Curry._1(fn, self$1);
-                }));
-          oldSelf.current = self$1;
-          return undefined;
-        }));
-  var mostRecentAllowedRender = React.useRef(React.useMemo((function () {
-              return Curry._1(componentSpec[/* render */9], self$1);
-            }), ([])));
-  if (hasBeenCalled.current && oldSelf.current[/* state */1] !== state && Curry._1(componentSpec[/* shouldUpdate */8], /* record */[
-          /* oldSelf */oldSelf.current,
-          /* newSelf */self$1
-        ])) {
-    Curry._1(componentSpec[/* willUpdate */7], /* record */[
-          /* oldSelf */oldSelf.current,
-          /* newSelf */self$1
-        ]);
-    mostRecentAllowedRender.current = Curry._1(componentSpec[/* render */9], self$1);
-  }
-  return mostRecentAllowedRender.current;
-}
-
-var Router = [
-  push,
-  replace,
-  watchUrl,
-  unwatchUrl,
-  url,
-  useUrl
-];
+var Router = 0;
 
 exports.statelessComponent = statelessComponent;
 exports.statelessComponentWithRetainedProps = statelessComponentWithRetainedProps;
@@ -676,5 +414,4 @@ exports.wrapReasonForJs = wrapReasonForJs;
 exports.createDomElement = createDomElement;
 exports.wrapJsForReason = wrapJsForReason;
 exports.Router = Router;
-exports.useRecordApi = useRecordApi;
 /* dummyInteropComponent Not a pure module */
