+++ lib/js/src/ReasonReactRouter.js
@@ -1,0 +1,164 @@
+'use strict';
+
+var Curry = require("bs-platform/lib/js/curry.js");
+var React = require("react");
+
+function safeMakeEvent(eventName) {
+  if (typeof Event === "function") {
+    return new Event(eventName);
+  } else {
+    var $$event = document.createEvent("Event");
+    $$event.initEvent(eventName, true, true);
+    return $$event;
+  }
+}
+
+function path() {
+  var match = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined) {
+    var raw = match.location.pathname;
+    switch (raw) {
+      case "" : 
+      case "/" : 
+          return /* [] */0;
+      default:
+        var raw$1 = raw.slice(1);
+        var match$1 = raw$1[raw$1.length - 1 | 0];
+        var raw$2 = match$1 === "/" ? raw$1.slice(0, -1) : raw$1;
+        var a = raw$2.split("/");
+        var _i = a.length - 1 | 0;
+        var _res = /* [] */0;
+        while(true) {
+          var res = _res;
+          var i = _i;
+          if (i < 0) {
+            return res;
+          } else {
+            _res = /* :: */[
+              a[i],
+              res
+            ];
+            _i = i - 1 | 0;
+            continue ;
+          }
+        };
+    }
+  } else {
+    return /* [] */0;
+  }
+}
+
+function hash() {
+  var match = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined) {
+    var raw = match.location.hash;
+    switch (raw) {
+      case "" : 
+      case "#" : 
+          return "";
+      default:
+        return raw.slice(1);
+    }
+  } else {
+    return "";
+  }
+}
+
+function search() {
+  var match = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined) {
+    var raw = match.location.search;
+    switch (raw) {
+      case "" : 
+      case "?" : 
+          return "";
+      default:
+        return raw.slice(1);
+    }
+  } else {
+    return "";
+  }
+}
+
+function push(path) {
+  var match = typeof (history) === "undefined" ? undefined : (history);
+  var match$1 = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined && match$1 !== undefined) {
+    match.pushState(null, "", path);
+    match$1.dispatchEvent(safeMakeEvent("popstate"));
+    return /* () */0;
+  } else {
+    return /* () */0;
+  }
+}
+
+function replace(path) {
+  var match = typeof (history) === "undefined" ? undefined : (history);
+  var match$1 = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined && match$1 !== undefined) {
+    match.replaceState(null, "", path);
+    match$1.dispatchEvent(safeMakeEvent("popstate"));
+    return /* () */0;
+  } else {
+    return /* () */0;
+  }
+}
+
+function url() {
+  return /* record */[
+          /* path */path(/* () */0),
+          /* hash */hash(/* () */0),
+          /* search */search(/* () */0)
+        ];
+}
+
+function watchUrl(callback) {
+  var match = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined) {
+    var watcherID = function () {
+      return Curry._1(callback, url(/* () */0));
+    };
+    match.addEventListener("popstate", watcherID);
+    return watcherID;
+  } else {
+    return (function () {
+        return /* () */0;
+      });
+  }
+}
+
+function unwatchUrl(watcherID) {
+  var match = typeof (window) === "undefined" ? undefined : (window);
+  if (match !== undefined) {
+    match.removeEventListener("popstate", watcherID);
+    return /* () */0;
+  } else {
+    return /* () */0;
+  }
+}
+
+function useUrl() {
+  var match = React.useState(url);
+  var setUrl = match[1];
+  React.useLayoutEffect((function () {
+          var watcherId = watchUrl((function (url) {
+                  return Curry._1(setUrl, (function () {
+                                return url;
+                              }));
+                }));
+          return (function () {
+                    return unwatchUrl(watcherId);
+                  });
+        }), ([]));
+  return match[0];
+}
+
+var dangerouslyGetInitialUrl = url;
+
+exports.push = push;
+exports.replace = replace;
+exports.watchUrl = watchUrl;
+exports.unwatchUrl = unwatchUrl;
+exports.dangerouslyGetInitialUrl = dangerouslyGetInitialUrl;
+exports.useUrl = useUrl;
+/* react Not a pure module */
