+++ src/ReasonReactCompat.rei
@@ -1,0 +1,5 @@
+open ReasonReact;
+
+let useRecordApi:
+  componentSpec('state, 'state, 'retainedProps, 'retainedProps, 'action) =>
+  reactElement;
